+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+### New features:
+
+- [] Add new built-in rules `unicodeIdStart` and `unicodeIdContinue`.
 - [] Add `experimentalApplySyntactic`, an experimental built-in rule to allow application of a syntactic rule from a lexical context (see #136 for details).
 
 ## v16.0.0 - Oct 1, 2021